We will address this issue one by one either in the branch or in the main trunk, depending if it requires a lot of surgery or not.

+++ edu.rice.cs.hpc.data/src/edu/rice/cs/hpc/data/experiment/metric/DerivedMetric.java
@@ -117,7 +117,7 @@ public class DerivedMetric extends BaseMetric {
 			dVal = objVal.doubleValue();
 		}
 		if(this.getPercent()){
-			return new MetricValue(dVal,dVal/this.dRootValue);
+			return new MetricValue(dVal, ((float) dVal/this.dRootValue));
 		} else {
 			return new MetricValue(dVal);
 		}
